requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cmdstanpy==0.9.5
coinapi.rest.v1==1.3
coloredlogs==15.0
convertdate==2.3.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
Django==3.1.6
django-bootstrap4==2.3.1
django-debug-toolbar==3.2
django-extensions==3.1.1
django-fontawesome-5==1.0.18
django-model-utils==4.1.1
django-pandas==0.6.4
ephem==3.7.7.1
fbprophet==0.7.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hijri-converter==2.1.1
holidays==0.10.5.2
humanfriendly==9.1
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
Markdown==3.3.3
matplotlib==3.3.4
nbformat==5.1.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
prompt-toolkit==3.0.16
protobuf==3.14.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.0
PyMeeus==0.3.7
pyparsing==2.4.7
pyrsistent==0.17.3
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
scikit-learn==0.24.1
scipy==1.6.0
setuptools-git==1.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.56.2
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
